+++ index.js
@@ -1,17 +1,12 @@
-const readline = require('readline-sync')
+const robots = {
+    userInput: require('./robots/user-inputs.js'),
+    text: require('./robots/text.js')
+ } 
 function start(){
  const content = {}
- content.searchTerm = askAndReturnSearchTerm()
- content.prefix = askAndReturnPrefix()
- function askAndReturnSearchTerm(){
-     return readline.question('Type a Wikipedia search term: ')
- }
- function askAndReturnPrefix(){
-    const prefixes = ['Who is', 'What is','The History of']
-    const selectedPrefixIndex = readline.keyInSelect(prefixes, 'Choose one option:')
-    const selectedPrefixText = prefixes[selectedPrefixIndex]
-    return selectedPrefixText
- }
-  console.log(content)
+
+ robots.userInput(content)
+ robots.text(content)
+ console.log(content)
 }
 start()
