+++ package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/raphaellc/video-maker#readme",
   "dependencies": {
+    "algorithmia": "^0.3.10",
     "readline-sync": "^1.4.10"
   }
 }
